+++ client/dashboard/utils/site-badge.ts
@@ -1,7 +1,16 @@
 import { isSitePlanTrial } from '../sites/plans';
 import { getSiteBlockingStatus } from './site-status';
 import { isP2 } from './site-types';
-import type { Site, SiteBadge } from '@automattic/api-core';
+import type { Site } from '@automattic/api-core';
+
+export type SiteBlockingStatus =
+	| 'deleted'
+	| 'migration_pending'
+	| 'migration_started'
+	| 'difm_lite_in_progress'
+	| null;
+
+export type SiteBadge = 'staging' | 'trial' | 'p2' | SiteBlockingStatus;
 
 export function getSiteBadge( site: Site ): SiteBadge {
 	const status = getSiteBlockingStatus( site );
